Rewrite 13/14: Lustre IO (#156)
This PR add the Lustre I/O classes.

## LFS wrapper
I've made a wrapper around calls to the `lfs(1)` binary in `alpenhorn.io.lfs.py`.  It is not a complete wrapper for that command and only able to run these commands:
* "lfs quota" to retrieve the group quota for a path.  The user group used must be specified in the I/O config.  Can also handle the quota max override needed for cedar's `/nearline`.
* "lfs hsm_state" to retrieve the HSM state for a file.  The state itself is represented with the `HSMState` enum and is one of:
  * `MISSING`: file is not present on disk or tape
  * `UNARCHIVED`: file exists on disk, but not on tape
  * `RESTORED`: file exists on both disk and tape
  * `RELEASED`: file exists on tape only
* "lfs hsm_restore" changes state `RELEASED -> RESTORED`
* "lfs hsm_release" changes state `RESTORED -> RELEASED`

## LustreQuota I/O

This is a very simple stepping-stone I/O module to Lustre HSM I/O.  It's exactly the same as DefaultIO, except is uses "lfs quota" to determine free space.  This behaviour re-creates the special casing we have in place in alpenhorn-1 for the "cedar_online" node.

## Lustre HSM I/O
The LustreHSM I/O is where all the HSM management occurs.  In general this is via updated versions of I/O methods from Default I/O. It also will automatically release files from disk (see `release_files`) as needed to keep enough headroom available on the HSM disk so that I/O can continue to happen.

The LustreHSM Group needs two nodes in it.  One of them must be a "LustreHSM" node, used as the "primary node", and another node, of any non-LustreHSM type, used as a "secondary node" for small files.  The group I/O config may specify the threshold for small files.