fix(io): Add missing BaseGroupIO.set_storage()
Also:
* renamed BaseNodeIO.update() to BaserNodeIO.set_storage().
* enforce keyword arguments in UpdateableGroup.reinit().

The "set_storage" methods are called at the top of every
update loop to replace the StorageNode/Group of the I/O classes.
"update" was a very poor choice for their name because there
are already too many things referred to as "update", and the
name was always just going to lead to confusion.

(The erstwhile "update()" method was always called _before_ the
"before_update()" method, which is a pretty clear indication there's
something wrong with the name.)