fix(update): Handle duplicate requests
If there are multiple pending requests for the same file to
the same group, alpenhorn should only ever be willing to process one
of them.

We take the first one we see, and ignore the rest.  (They'll get
dealt with later, once the one we're working on has been resolved.)