fix(run_command): fix subprocess timeout recovery
Normally, when a transfer times out, it's because
it's stuck deep down in I/O wait, meaning SIGKILL
is not going to be handled in a timely manner,
and, so, trying to `.communicate()` with it again
is just futile.

With the current code, after a timeout, the worker
just gets stuck in the same wait it tried to get
out of with the kill, as if there were no timeout
given.

So, let's just send it a kill and then abandon it.

Related to #175 (abandonning subprocesses may lead
to zombies), but this doesn't really address
the items brought up in that issue, which are higher-
level ideas.