Drop globre from test requirements
Back in #163 I was content to solve #160 by demoting `globre`
from a requirement to a test requirement, but I think even here it's
an unnecessary complication for something we're not using in
production.

This rewrites the `pattern_importer` example extension to accept only
regular expressions and updates the patterns in `tests/test_service.py`
to follow.

Also updated how the `pattern_importer` is imported in the
`test_service` test to handle cases where the current directory is not
in the import path or the test suite isn't run from the top-level
alpenhorn directory.

Closes #131