fix(lustrehsm): Only call hsm_restore once during restore-wait
Calling it more than once seems to result in the second call not
completing until the file is restored (locking up the worker).

Probably we should add a timeout to the restore-wait (#188) but I
want to run it as-is for a while to see how this goes.

Also, fixed the `check_async` call.

Also, some logging tweaks.