fix(update): Remove I/O from main loop
This fixes another instance of unnecessary I/O in the
MainThread.  Specifically, when processing a pull request,
the UpdateableGroup was searching the group's node(s) for
an unknown file on disk.

This check has been moved to a Task (`_default_asyncs.group_search_async`)
and is now done later in the pull request processing, after non-I/O
checks have finished (which should, overall, reduce the amount of I/O
going on).

The tricky part here, is that this is the first Task that needs to run
at the Group level, instead of the Node level, so some work has been
done to add Task support to groups (which was always my plan, it just
turned out there were no group tasks when I refactored alpenhorn).

Has been deployed to cedar