fix(io/lustrehsm): Handle timeout in hsm_restore call.
On `cedar`, at least, running `lfs hsm_restore` on a file that
is already being restored causes `lfs` to wait for a very long
time (~15 minutes) before returning.

This deals with that eventuality by returning a timeout error if
the hsm_restore call takes more than a minute.  On timeout, alpenhorn
will assume success (the idea being the timeout was due to the file
restore being already in progress).

However, to guard against a false assumption, alpenhorn will retry
the restore request after an hour.  This retry is now also done (but
after four hours) on a successful hsm_restore request.  This is to
mitigate a potential issue I've been thinking about: alpenhorn
submitting a successful hsm_restore request that is later abandonned
by the HSM backend before the file is restored, which has been a
potential problem since #189.