refactor: reorganize modules
Separate server-only, client-only, db, scheduler, common.  Also
re-organised the test suite to follow suit.

This is a first step towards resuscitating the client (which isn't
meant to be a second-class part of alpenhorn).  Some further tweaks
may be necessary.

For both the db and scheduler modules, the typically used symbols
are imported into __init__ so that you can simply import them from
the base module, rather than having to know which submodule to use.

Updated import of table models in the client as part of the client prep
work.