fix(client): Fix "acq" command group (#206)
There are only three commands here now:

- list: to list acquisition
- files: to list acquisition files
- show: to show an acquisition

The "where" command is now handled by "show":
```
  alpenhorn acq show ACQNAME --show-nodes
```

The "syncable" command will be re-implemented as part of the "syncable"
command in the "file" command group, where an acq-name limit can be
specified. Something like:
```
  alpenhorn file syncable --acq=ACQ
```

The `pyproject.toml` change is to allow filenames in the test suite with
the same name (e.g. `test_list.py`)