fix(db): Fix db.threadsafe (#212)
Something I found while testing #210:

If we want db._base.threadsafe to be imported into `__init__`, it can't
be a scalar (because it's going to change after the import).

Instead, make it a function.

Bonus: no need to try to import `subprocess` every time a command is
run. Just import it once at module import.