feat(server): A few optimations for deep queues (#219)
A couple of changes to reduce the amount of churn when it's taking a
long time between job submission and execution because of a lot of other
stuff already in the queue:

* when looking for file to ready for a remote pull, don't (re-)submit a
ready job for things that are already ready (which is harmless to do,
but not in any way helpful

* at the start of a pull job, re-check whether the file copy already
exists on the destination node. Again: would be harmless not to check: a
new copy of the same file would overwrite the old one, but there's no
harm to checking.

This takes care of the case when two pulls for the same file happen in
series. It does not solve the case of the two identical pulls occurring
in parallel.