refactor: Use "daemon" and "cli" instead of "server/client" (#220)
Talking to the CGEM folks, I think our (my) use of the terms "server" to
refer to the long-running process that manages alpenhorn nodes, and
"client" to refer to the command-line interface to the data index are
unnecessarily confusing.

I want to move to using "daemon" and "cli" for these pieces of alpenhorn
instead.

The terms "server" and "client" imply that the two pieces of alpenhorn
are a normal server-client system, which they are not. To over-simplify
things, normally a "server" is a long-running process that produces data
for users. The users use the "client" to connect to the server and
retrieve the data.

The two parts of alpenhorn do neither of these things. The daemon is not
at all user-facing. It doesn't "serve" anything that anyone could
retrieve. It's concerned with managing files on disk (or tape).

Neither is the alpenhorn CLI a "client" in the traditional sense. It
doesn't interact with the server directly at all, and there's no API or
protocol defined for communication between the two parts. The CLI is a
tool for reading and writing the data index. By design, the two pieces
can only communicate indrectly through the database.

AFAICT, the alpenhorn daemon was originally called a "daemon" (that's
why it's `alpenhornd`. Later this migrated to "service" (the alpenhorn-1
daemon is implemented in the module `alpenhorn.service`) and then later
still to "server". A lot of the "server" nomenclature shows up during my
multithreaded rewrite of the daemon. So, using "daemon" is something of
a return to the original name.

The CLI has been referred to as both the "CLI" and the "client"
interchangably since we started. (alpenhorn-1's CLI is implemented in
the module `alpenhorn.client`). Here I want to push "CLI" over "client"
to avoid slipping back into "client-server" terms in the future.

Though, I am willing to listen to other options for the names of these
two pieces. At some point we need to write some documentation for this
system, and it would be good to have some concrete terms to work with
for that.