feat(CLI): StorageTransferAction support
This is based on my suggestion to remove the `action` group in #202,
which I think was ill-defined.

Instead:

* Adds an `--actions` flag to `node show` and `group show` to
  show relevant `StorageTransferAction`s affecting the node/group.
* Adds an `--all` flag to `node show` and `group show` to provide a
  way to list everything without having to know what lists what.
* Adds two new commands to manipulate `StorageTransferAction`s:
  * `group auto-sync GROUP NODE` to set/clear the autosync flag
  * `node auto-clean NODE GROUP` to set/clear the autoclean flag

For a given (NODE, GROUP) pair, these two commands modify the same row
in the `StorageTransferAction` table (the pair has to be unique in the
table).  So, there's ambiguity over which command goes with which "noun".
The way I've done it, the command is associated with the object that the
auto-action modifies (auto-sync adds files to GROUP; auto-clean deletes
files from NODE).  There's probably an argument for doing it the other
way; I'm not sure it matters much, so long as we are consistent about
it.

Also: should the command names have a dash in them?  Or would `group
autosync GROUP NODE` etc. be better?