fix(CLI): fix "db init"
Not much to fix here.  The extension tables aren't handled but
they can't be unless we were to create some sort of new extension
type which provides a list of tables to create.

Given that the only use of such an extension would be to pass them
to `create_tables` in this function, it may be more trouble than
it's worth.  Probably easier for both alpenhorn and the extension
creators to just tell them to create their own tables.

In any case, it's very low priority.