fix(CLI): Re-implement "node verify" (#229)
This is a re-implementation of "node verify" which removes all I/O from
the CLI.

As previously discussed, I think it's difficult to have I/O in the
client. In general, it's difficult to keep I/O clean unless we're
very specific on how I/O works, which we've become less-so with the new
I/O framework stuff. So, keeping I/O to the daemon vastly simplifies
things.

Even with the preexisting simplified version of verification that I'm
replacing, issues like #141 show how easy it is to get in to situations
where verification fails. The verify-in-cli scheme requires both running
where the node is hosted and from a user that has access to the files on
the node.

I don't know how we would ever run a CLI-based verify on a node like
`cedar_nearline` where it would have to wait for files to be recalled.

This new verify command can also be used to "cancel" a verification,
which really means force the `has_file='M'` suspect file into one of the
other `has_file` values.

I've taken the check-confirm-update stuff from my implementation of
`clean` and put it in a utility function so it can be used elsewhere
(like here).

With this PR, I think neither of the issues linked to the verify command
are relevant anymore:

* With the removal of I/O from the CLI, where its run is now immaterial:
#141
* The underlying problem with #133 was trying to find a way to check
files on Storage Nodes hosted where there was no daemon running. In
alpenhorn1 you could do some of the daemon's work in the client, but it
was never a complete solution, and in practice we always ended up
putting daemons on machines where we had nodes (e.g. the transport disk
off-load machine here at UBC), even if that daemon spent most of its
time doing nothing.

The current answer to #133 would be: run the daemon to maintain the
node.

There are also a couple of commits in this PR tidying up minor things I
overlooked in the previous PRs.